Update UserEntity.java
This file contains the basic set up for the user entity class. Using Lombok import, the boilerplate code of getters and setters are eliminated using annotations (@Id, @Getter, @Setter). The fields are automatically structured to be used in future programming of this application. 

@oneToMany specifies the ownership of the User Entity in a bilateral relationship. Starting with the 'User side'.

(mapepdBy) - Talks to JPA/Hibernate and tells it how to create instances of the UserEntity class.